requirements.txt
adrf==0.1.6
asgiref==3.8.1
async-property==0.2.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==5.0.6
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
filelock==3.15.4
fsspec==2024.6.0
huggingface-hub==0.23.4
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
numpy==2.0.0
packaging==24.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rpds-py==0.18.1
safetensors==0.4.3
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wheel==0.43.0
zope.interface==6.4.post2